Project | Feed forward working. Running too slowly

+++ project/FishRecord.py
@@ -18,7 +18,7 @@ class FishRecord:
         trajectory_str, fish_str = parts[0].split('_')
         self.trajectory_id = int(trajectory_str)
         self.fish_id = int(fish_str)
-        self.species = torch.nn.functional.one_hot(torch.tensor(index), num_classes=23)
+        self.species = torch.nn.functional.one_hot(torch.tensor(index), num_classes=23).float()
         self.file_path = f"{root_path}/{prefix}_image/{prefix}_{parts[1].zfill(2)}/{prefix}_{trajectory_str.zfill(12)}_{fish_str}.png"
